backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────
fastapi==0.95.2
uvicorn[standard]==0.27.1
starlette==0.27.0
python-multipart==0.0.9

# ── Data Validation ───────────────────────────────────────
pydantic==1.10.13

# ── Database ──────────────────────────────────────────────
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# ── Authentication ────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ── ML / Data Science ─────────────────────────────────────
numpy==1.26.4
joblib==1.3.2
scikit-learn==1.4.0

# ── AI / External APIs ────────────────────────────────────
google-generativeai==0.8.5

# ── Optional ML Runtime ───────────────────────────────────
onnxruntime==1.17.0

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1